Revert "Plan 9 from Bell Labs Fourth Edition 2007-02-23"
This reverts commit 41fb754a868aff65c81b63b0b07ae1fab2010fd9.

+++ sys/src/cmd/cfs/disk.h
@@ -6,8 +6,8 @@ typedef struct Disk	Disk;
 struct Disk
 {
 	Bcache;
-	ulong	nb;	/* number of blocks */
-	ulong	nab;	/* number of allocation blocks */
+	int	nb;	/* number of blocks */
+	int	nab;	/* number of allocation blocks */
 	int	b2b;	/* allocation bits to a block */
 	int	p2b;	/* Dptr's per page */
 	char	name[KNAMELEN];
